check post-like

+++ src/main/java/com/zetavn/api/service/impl/PostLikeServiceImpl.java
@@ -17,7 +17,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.io.NotActiveException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -51,12 +50,31 @@ public class PostLikeServiceImpl implements PostLikeService {
             return ApiResponse.error(HttpStatus.NOT_FOUND, e.getMessage());
         }
         }
+        @Override
+     public ApiResponse<?> checkPostLike(String userId, String postId)  {
+        try{
+            PostEntity post = postRepository.findById(postId).orElseThrow(NullPointerException::new);
+            UserEntity user = userRepository.findById(userId).orElseThrow(NullPointerException::new);
+            PostLikeEntity postLikeCheck = postLikeRepository.findPostLikeEntityByPostEntityAndUserEntity(post,user);
+            if(postLikeCheck==null){
+                return  ApiResponse.success(HttpStatus.OK,"User not like",false);
+            }else {
+                return  ApiResponse.success(HttpStatus.OK,"User liked",true);
+            }
+        }catch (Exception e){
+            return  ApiResponse.error(HttpStatus.BAD_REQUEST,"Error",e.getMessage());
+        }
+     }
 
     @Override
     public ApiResponse<?> createPostLike(PostLikeRequest postLikeRequest) {
         try{
             PostEntity post = postRepository.findById(postLikeRequest.getPostId()).orElseThrow(NullPointerException::new);
             UserEntity user = userRepository.findById(postLikeRequest.getUserId()).orElseThrow(NullPointerException::new);
+            PostLikeEntity postLikeCheck = postLikeRepository.findPostLikeEntityByPostEntityAndUserEntity(post,user);
+            if(postLikeCheck!=null){
+                return  ApiResponse.error(HttpStatus.BAD_REQUEST,"User liked",new PostLikeResponse(user.getUserId(),post.getPostId()));
+            }
             PostLikeEntity postLike = new PostLikeEntity();
             postLike.setPostEntity(post);
             postLike.setUserEntity(user);
@@ -73,6 +91,9 @@ public class PostLikeServiceImpl implements PostLikeService {
             PostEntity post = postRepository.findById(postLikeRequest.getPostId()).orElseThrow(NullPointerException::new);
             UserEntity user = userRepository.findById(postLikeRequest.getUserId()).orElseThrow(NullPointerException::new);
             PostLikeEntity postLike = postLikeRepository.findPostLikeEntityByPostEntityAndUserEntity(post,user);
+            if(postLike==null){
+                return  ApiResponse.error(HttpStatus.BAD_REQUEST,"User not like",new PostLikeResponse(user.getUserId(),post.getPostId()));
+            }
             postLikeRepository.delete(postLike);
             return ApiResponse.success(HttpStatus.OK,"Delete post-like success",null);
